+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/apache/ApacheRestCallFactory.java
@@ -20,6 +20,7 @@ import at.porscheinformatik.happyrest.RestUtils;
 import at.porscheinformatik.happyrest.jackson.JacksonBasedFormatter;
 import at.porscheinformatik.happyrest.jackson.JacksonBasedParser;
 import at.porscheinformatik.happyrest.util.ByteArrayParser;
+import at.porscheinformatik.happyrest.util.CharArrayParser;
 import at.porscheinformatik.happyrest.util.StringParser;
 import at.porscheinformatik.happyrest.util.TextPlainFormatter;
 
@@ -34,7 +35,9 @@ public class ApacheRestCallFactory implements RestCallFactory
     public static ApacheRestCallFactory create(RestLoggerAdapter loggerAdapter, ObjectMapper mapper)
     {
         RestFormatter formatter = RestFormatter.of(new JacksonBasedFormatter(mapper), new TextPlainFormatter());
-        RestParser parser = RestParser.of(new JacksonBasedParser(mapper), StringParser.INSTANCE, ByteArrayParser.INSTANCE);
+        RestParser parser = RestParser
+            .of(new JacksonBasedParser(mapper), StringParser.INSTANCE, CharArrayParser.INSTANCE,
+                ByteArrayParser.INSTANCE);
 
         return new ApacheRestCallFactory(loggerAdapter, formatter, parser);
     }
@@ -50,7 +53,8 @@ public class ApacheRestCallFactory implements RestCallFactory
 
     public ApacheRestCallFactory(RestLoggerAdapter loggerAdapter, RestFormatter formatter, RestParser parser)
     {
-        this(loggerAdapter, formatter, parser, ProxySelector.getDefault(), null, RestUtils.getUserAgent("Apache's HttpClient"));
+        this(loggerAdapter, formatter, parser, ProxySelector.getDefault(), null,
+            RestUtils.getUserAgent("Apache's HttpClient"));
     }
 
     public ApacheRestCallFactory(RestLoggerAdapter loggerAdapter, RestFormatter formatter, RestParser parser,