+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/util/StringParser.java
@@ -5,6 +5,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 
 import at.porscheinformatik.happyrest.GenericType;
 import at.porscheinformatik.happyrest.RestParser;
@@ -30,9 +31,9 @@ public class StringParser implements RestParser
     }
 
     @Override
-    public <T> Object parse(String contentType, GenericType<?> type, InputStream in) throws RestParserException
+    public <T> String parse(String contentType, GenericType<?> type, InputStream in) throws RestParserException
     {
-        Charset charset = RestUtils.extractContentTypeCharset(contentType);
+        Charset charset = RestUtils.extractContentTypeCharset(contentType, StandardCharsets.UTF_8);
 
         try (Reader reader = new InputStreamReader(in, charset))
         {